projects/Recipes/requirements.txt
pillow
pytesseract
opencv-python
numpy
jinja2
# Development dependencies
flake8
black
isort
pytest
